// frontend/src/api/models/CreateTherapistDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime'
/**
 *
 * @export
 * @interface CreateTherapistDTO
 */
export interface CreateTherapistDTO {
  /**
   *
   * @type {string}
   * @memberof CreateTherapistDTO
   */
  email?: string
  /**
   *
   * @type {string}
   * @memberof CreateTherapistDTO
   */
  password?: string
}

/**
 * Check if a given object implements the CreateTherapistDTO interface.
 */
export function instanceOfCreateTherapistDTO(value: object): value is CreateTherapistDTO {
  return true
}

export function CreateTherapistDTOFromJSON(json: any): CreateTherapistDTO {
  return CreateTherapistDTOFromJSONTyped(json, false)
}

export function CreateTherapistDTOFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean
): CreateTherapistDTO {
  if (json == null) {
    return json
  }
  return {
    email: json['email'] == null ? undefined : json['email'],
    password: json['password'] == null ? undefined : json['password'],
  }
}

export function CreateTherapistDTOToJSON(json: any): CreateTherapistDTO {
  return CreateTherapistDTOToJSONTyped(json, false)
}

export function CreateTherapistDTOToJSONTyped(
  value?: CreateTherapistDTO | null,
  ignoreDiscriminator: boolean = false
): any {
  if (value == null) {
    return value
  }

  return {
    email: value['email'],
    password: value['password'],
  }
}